backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.6.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
websockets>=12.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1

# AI/ML - Hugging Face
huggingface-hub==0.29.1
gradio-client==2.0.1

# Image Processing
pillow==10.2.0

# Optional: Local LLM Support
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
torch>=2.1.0

# V2 Features - Vector Memory
sentence-transformers>=2.2.0
chromadb>=0.4.0

# V2 Features - Better async
aiohttp>=3.9.0
